skip some additional config options when restoring
-unusedX in all cases, even when creating
-parent when merging config from backup

+++ src/PVE/LXC/Create.pm
@@ -154,8 +154,9 @@ sub restore_and_configure {
 	    my $oldconf = PVE::LXC::Config::parse_pct_config("/lxc/$vmid.conf", $raw);
 
 	    foreach my $key (keys %$oldconf) {
-		next if $key eq 'digest' || $key eq 'rootfs' || $key eq 'snapshots' || $key eq 'unprivileged';
+		next if $key eq 'digest' || $key eq 'rootfs' || $key eq 'snapshots' || $key eq 'unprivileged' || $key eq 'parent';
 		next if $key =~ /^mp\d+$/; # don't recover mountpoints
+		next if $key =~ /^unused\d+$/; # don't recover unused disks
 		$conf->{$key} = $oldconf->{$key} if !defined($conf->{$key});
 	    }
 	    unlink($pct_cfg_fn);
